feat(api): Add get_current_admin dependency and restructure user routers
