+++ services/api/src/users/routers/__init__.py
@@ -1,0 +1,4 @@
+from src.users.routers.admin import router as admin_router
+from src.users.routers.profile import router as profile_router
+
+__all__ = ["admin_router", "profile_router"]
