+++ services/api/src/users/routers/admin.py
@@ -1,14 +1,14 @@
 from fastapi import APIRouter, Depends, status
-from src.core.dependencies import DatabaseDep, CurrentUser
+from src.core.dependencies import DatabaseDep, get_current_admin
 from src.core.responses import ApiResponse
-from src.core.exceptions import Forbidden
-from src.users.schemas import UserCreate, AdminUserUpdate, ProfileUpdate, UserResponse
+from src.users.schemas import UserCreate, AdminUserUpdate, UserResponse
 from src.users.service import UserService
 
 
 router = APIRouter(
     prefix="/users",
     tags=["Users"],
+    dependencies=[Depends(get_current_admin)],  # All routes require admin by default
 )
 
 
@@ -23,7 +23,6 @@ def get_user_service(db: DatabaseDep) -> UserService:
     description="Retrieve a list of all users in the system",
 )
 async def get_all_users(
-    current_user: CurrentUser,
     service: UserService = Depends(get_user_service),
 ) -> ApiResponse[list[UserResponse]]:
     """
@@ -38,47 +37,6 @@ async def get_all_users(
 
 
 @router.get(
-    "/me",
-    response_model=ApiResponse[UserResponse],
-    summary="Get my profile",
-    description="Retrieve the user's own profile info.",
-)
-async def get_my_profile(
-    current_user: CurrentUser,
-) -> ApiResponse[UserResponse]:
-    """
-    GET /users/me
-    """
-    return ApiResponse(
-        success=True,
-        message="Profile retrieved successfully",
-        data=current_user,
-    )
-
-
-@router.put(
-    "/me",
-    response_model=ApiResponse[UserResponse],
-    summary="Update my profile",
-    description="Update your own profile info.",
-)
-async def update_my_profile(
-    profile_data: ProfileUpdate,
-    current_user: CurrentUser,
-    service: UserService = Depends(get_user_service),
-) -> ApiResponse[UserResponse]:
-    """
-    PUT /users/me
-    """
-    updated_user = await service.update_profile(current_user.id, profile_data)
-    return ApiResponse(
-        success=True,
-        message="Profile updated successfully",
-        data=updated_user,
-    )
-
-
-@router.get(
     "/{user_id}",
     response_model=ApiResponse[UserResponse],
     summary="Admin gets user by ID",
@@ -86,19 +44,12 @@ async def update_my_profile(
 )
 async def get_user_by_id(
     user_id: int,
-    current_user: CurrentUser,
     service: UserService = Depends(get_user_service),
 ) -> ApiResponse[UserResponse]:
     """
     GET /users/{user_id}
     Admin-only endpoint.
-
-    Note: this explicit role checking is just for the mvp but in the future,
-    this better be replaced with a dependency for RBAC
     """
-    # Simple role check
-    if current_user.role != "admin":
-        raise Forbidden(detail="Only admins can access this endpoint")
 
     user = await service.get_user_by_id(user_id)
     return ApiResponse(
@@ -117,7 +68,6 @@ async def get_user_by_id(
 )
 async def create_user(
     user_data: UserCreate,
-    current_user: CurrentUser,
     service: UserService = Depends(get_user_service),
 ) -> ApiResponse[UserResponse]:
     """
@@ -140,19 +90,12 @@ async def create_user(
 async def update_user(
     user_id: int,
     user_data: AdminUserUpdate,
-    current_user: CurrentUser,
     service: UserService = Depends(get_user_service),
 ) -> ApiResponse[UserResponse]:
     """
     PUT /users/{user_id}
     Admin-only endpoint.
-
-    Note: this explicit role checking is just for the mvp but in the future,
-    this better be replaced with a dependency for RBAC
     """
-    # Simple role check
-    if current_user.role != "admin":
-        raise Forbidden(detail="Only admins can access this endpoint")
 
     updated_user = await service.admin_update_user(user_id, user_data)
     return ApiResponse(
@@ -170,7 +113,6 @@ async def update_user(
 )
 async def delete_user(
     user_id: int,
-    current_user: CurrentUser,
     service: UserService = Depends(get_user_service),
 ) -> None:
     """
